fix: move LanguageInfo schemas to backend/schemas
